requirements.txt
streamlit
yfinance
google-generativeai
GoogleNews
praw
pypdf
youtube-transcript-api
python-dotenv
pandas
plotly
